Correcting GET/POST routines for providing response details
The GET and POST handling routines now only return response details if
there were any -- in other words when the level of detail is provided in
a successful HTTP transaction even if it was an error transaction.  If
an error occurs at a lower level where no response was provided from the
server, we simply allow that to throw up directly.