Cleaning up the DNS challenge tests and adding HTTP tests
* Cleaning up the process for generating DNS tests and submitting them
for verification.
* The DNS tests are passing successfully against our private local
Boulder CA instance, but unfortunately failing against the public STAGE
Boulder CA instance; we have confirmation that it's pby because the DNS
challenge verification is not fully baked and is not going to be
supported for the Boulder 1.0 release:
https://github.com/letsencrypt/boulder/issues/720
* Adding fully fleshed out HTTP tests (simpleHttp challenge method) --
this is testing successfully against *both* private local and public
STAGE Boulder CA instances; in both cases we can successfully complete
the process through certificate request and issue
* Removing stand-alone cert request -- not a well-defined test without
the pre-requ of generating a new CSR each time.