Adding support for *well-known* service URLs and LE BETA/PROD endpoint
As per #34, I'm introducing the concept of a *well-known service" which
are friendly-labeled ACME service endpoints.  To start with, we have LE
BETA/PROD and LE STAGE.  The BETA/PROD will now become the default.

Additionally, a custom URI can still be specified as a Base Service URI
instead of a well-known label.