Report no identifiers even when there is no identifierRef
In the case that an identifier alias is used an exception will be reported if there are no identifiers.  If no alias is used, an attempt will be made to use the vault identifiers collection even if it is not yet assigned, leading to a null reference exception.  It seems the test to check for existing identifiers performed when an alias is used is just as valid when no identifier is used. This change moves the test so it is used in both scenarios.